fix: correctly handle ubisys device setup
Our implementation was working by accident, this is now corrected and we can write larger payloads.
This is done in the format Ubisys expects, however I did remove the old fallback code. It's complicated as, people should really update the firmware if they are running such and old version.

There are so many fw bug fixes missing if they don't, I think it's OK to remove this. There is now a warning logged telling people to upgrade.

Co-authored-by: Nerivec <62446222+Nerivec@users.noreply.github.com>