fix: LiXee ZLinky: describe actual default polling interval
The default description ("This device does not support reporting
electric measurements so it is polled instead. The default poll interval
is 60 seconds, set to -1 to disable.") mentioned a default of 60 seconds
but it is actually 600.
It was also misleading: not all attributes are polled, several of them
have reporting configured and are automatically updated.