fix: LiXee ZLinky: handle empty TIC command list
If the user defines a value for tic_command_whitelist and then clears
it, it remains in the configuration as an empty string.
An empty string should be the same as no value at all. Before this
change, when the string was empty, no attributes were polled.

Let's also suggest a Z2M restart in the description, it's the easiest way to
restart the poll.