Update konke.ts
Add support for older versions of the water sensor to enable the older KK-WA-J01W water sensor to display battery level.
The model 3AFE12010402102A differs from the other two models, which are KK-WA-J01W water sensors. The firmware of 3AFE12010402102A was written in 2020 and does not have the function to report battery level by itself. Therefore, this update adds a calculation function to enable Z2M to calculate the battery level of the water sensor through the battery voltage.
Author: KleinerSource, this is the external converter file he provided.